\name{mailminer-package}
\alias{mailminer-package}
\alias{mailminer}
\docType{package}
\title{
MailMiner
}
\description{
Helper functions for the MailMiner utility, a set of scripts for mining your emails, first used for chapter 6 in Everyday Programming
}
\details{
\tabular{ll}{
Package: \tab mailminer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-09-19\cr
License: \tab Apache 2.0\cr
LazyLoad: \tab yes\cr
}
Time and date frequency functions
}
\author{
Chang Sau Sheong

Maintainer: Chang Sau Sheong <sausheong@gmail.com>
}
\references{
None
}

\keyword{ package }
\seealso{
}
\examples{

}
